backend/requirements.txt
fastapi==0.111.0
uvicorn==0.41.0
supabase==2.5.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.2.1
python-multipart==0.0.9
